1. role added on all routes 2. using id rather than roleid and lastrowid

+++ backend/controller/appointment/appointmentStatusController.js
@@ -1,8 +1,13 @@
 const sqlite3 = require('sqlite3').verbose();
 const db = new sqlite3.Database('demo.db');
-const uuidv4 = require('uuid/v4');
+const {getGrants} = require('../../utitlity/roleManager');
 
 exports.deleteAppointmentStatus = async (req, res) => {
+  const permission1 = getGrants.can(req.user.role).deleteAny("appointmentStatus");
+  if (!permission1.granted) {
+    return res.status(403).json("Access forbidden for " + req.user.role);
+  }
+
   await db.run("DELETE from AppointmentStatus WHERE id= ?", req.params['id'], function (err) {
     if (err) {
       res.json(err.message).status(404);
@@ -10,29 +15,50 @@ exports.deleteAppointmentStatus = async (req, res) => {
       res.json(this.changes);
     }
   });
+
+
 };
 
 exports.saveAppointmentStatus = async (req, res) => {
+  const permission = getGrants.can(req.user.role).createAny("appointmentStatus");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden for " + req.user.role);
+  }
+
   const newAppointmentStatus = {
-    id: uuidv4(),
     status: req.body.status
   };
-  await db.run("INSERT INTO AppointmentStatus(id,status) VALUES(?,?)", [newAppointmentStatus.id, newAppointmentStatus.status], function (err) {
+
+  await db.run("INSERT INTO AppointmentStatus(status) VALUES(?)", [newAppointmentStatus.status], function (err) {
     if (err) {
       return res.status(400).send(err.message);
     } else {
       return res.json(newAppointmentStatus);
     }
   });
+
+
 };
 
 exports.getAllAppointmentStatus = async (req, res) => {
+  const permission1 = getGrants.can(req.user.role).readAny("appointmentStatus");
+  if (!permission1.granted) {
+    return res.status(403).json("Access forbidden for " + req.user.role);
+  }
+
   await db.all("SELECT * FROM appointmentStatus", function (err, rows) {
     return res.json(rows);
   });
+
 };
 
 exports.updateAppointmentStatus = async (req, res) => {
+
+  const permission1 = getGrants.can(req.user.role).updateAny("appointmentStatus");
+  if (!permission1.granted) {
+    return res.status(403).json("Access forbidden for " + req.user.role);
+  }
+
   await db.run("update AppointmentStatus set Status=? where id=?", [req.body.status, req.params["id"]], function (err) {
     if (err) {
       return res.status(400).send(err.message);
@@ -40,4 +66,6 @@ exports.updateAppointmentStatus = async (req, res) => {
       return res.json(this.changes);
     }
   });
+
+
 };
