+++ backend/controller/appointment/appointmentTypeController.js
@@ -1,8 +1,13 @@
 const sqlite3 = require('sqlite3').verbose();
 const db = new sqlite3.Database('demo.db');
-const uuidv4 = require('uuid/v4');
+const {getGrants} = require('../../utitlity/roleManager');
 
 exports.deleteAppointmentType = async (req, res) => {
+  const permission1 = getGrants.can(req.user.role).deleteAny("appointmentType");
+  if (!permission1.granted) {
+    return res.status(403).json("Access forbidden for " + req.user.role);
+  }
+
   await db.run("DELETE from AppointmentType WHERE id= ?", req.params['id'], function (err) {
     if (err) {
       res.json(err.message).status(404);
@@ -13,8 +18,12 @@ exports.deleteAppointmentType = async (req, res) => {
 };
 
 exports.saveAppointmentType = async (req, res) => {
+  const permission = getGrants.can(req.user.role).createAny("appointmentType");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden for " + req.user.role);
+  }
+
   const newAppointmentType = {
-    id: uuidv4(),
     type: req.body.type
   };
 
@@ -29,12 +38,22 @@ exports.saveAppointmentType = async (req, res) => {
 };
 
 exports.getAllAppointmentTypes = async (req, res) => {
+  const permission1 = getGrants.can(req.user.role).readAny("appointmentType");
+  if (!permission1.granted) {
+    return res.status(403).json("Access forbidden for " + req.user.role);
+  }
+
   await db.all("SELECT * FROM appointmentType", function (err, rows) {
     return res.json(rows);
   });
 };
 
 exports.updateAppointmentType = async (req, res) => {
+  const permission1 = getGrants.can(req.user.role).updateAny("appointmentType");
+  if (!permission1.granted) {
+    return res.status(403).json("Access forbidden for " + req.user.role);
+  }
+
   await db.run("update AppointmentType set type=? where id=?", [req.body.type, req.params["id"]], function (err) {
     if (err) {
       return res.status(400).send(err.message);
