+++ backend/controller/emergencyTitleController.js
@@ -1,8 +1,14 @@
 const sqlite3 = require('sqlite3').verbose();
 const db = new sqlite3.Database('demo.db');
-const uuidv4 = require('uuid/v4');
+const {getGrants} = require('../utitlity/roleManager');
 
 exports.deleteEmergencyTitle = async (req, res) => {
+
+  const permission = getGrants.can(req.user.role).deleteAny("emergencyTitle");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden " + req.user.role);
+  }
+
   await db.run("DELETE from EmergencyTitle WHERE id= ?", req.params['id'], function (err) {
     if (err) {
       res.json(err.message).status(404);
@@ -13,13 +19,17 @@ exports.deleteEmergencyTitle = async (req, res) => {
 };
 
 exports.saveEmergencyTitle = async (req, res) => {
+
+  const permission = getGrants.can(req.user.role).createAny("emergencyTitle");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden " + req.user.role);
+  }
+
   const newEmergencyTitle = {
-    id: uuidv4(),
     title: req.body.title
   };
 
-
-  await db.run("INSERT INTO EmergencyTitle(id,title) VALUES(?,?)", [newEmergencyTitle.id, newEmergencyTitle.title],
+  await db.run("INSERT INTO EmergencyTitle(title) VALUES(?)", [newEmergencyTitle.title],
     function (err) {
       if (err) {
         return res.status(400).send(err.message);
@@ -30,12 +40,24 @@ exports.saveEmergencyTitle = async (req, res) => {
 };
 
 exports.getAllEmergencyTitle = async (req, res) => {
+
+  const permission = getGrants.can(req.user.role).readAny("emergencyTitle");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden " + req.user.role);
+  }
+
   await db.all("SELECT * FROM EmergencyTitle", function (err, rows) {
     return res.json(rows);
   });
 };
 
 exports.updateEmergencyTitle = async (req, res) => {
+
+  const permission = getGrants.can(req.user.role).updateAny("emergencyTitle");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden " + req.user.role);
+  }
+
   await db.run("Update EmergencyTitle set title=? where id=?", [req.body.title, req.params["id"]], function (err) {
     if (err) {
       return res.status(400).send(err.message);
