+++ backend/controller/patient/patientTitleController.js
@@ -1,8 +1,14 @@
 const sqlite3 = require('sqlite3').verbose();
 const db = new sqlite3.Database('demo.db');
-const uuidv4 = require('uuid/v4');
+const {getGrants} = require('../../utitlity/roleManager');
 
 exports.deletePatientTitle = async (req, res) => {
+
+  const permission = getGrants.can(req.user.role).deleteAny("patientTitle");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden " + req.user.role);
+  }
+
   await db.run("DELETE from PatientTitle WHERE id= ?", req.params['id'], function (err) {
     if (err) {
       res.json(err.message).status(404);
@@ -13,12 +19,17 @@ exports.deletePatientTitle = async (req, res) => {
 };
 
 exports.savePatientTitle = async (req, res) => {
+
+  const permission = getGrants.can(req.user.role).createAny("patientTitle");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden " + req.user.role);
+  }
+
   const newPatientTitle = {
-    id: uuidv4(),
     title: req.body.title
   };
 
-  await db.run("INSERT INTO PatientTitle(id,title) VALUES(?,?)", [newPatientTitle.id, newPatientTitle.title], function (err) {
+  await db.run("INSERT INTO PatientTitle(title) VALUES(?)", [newPatientTitle.title], function (err) {
     if (err) {
       return res.status(400).send(err.message);
     } else {
@@ -28,12 +39,23 @@ exports.savePatientTitle = async (req, res) => {
 };
 
 exports.getAllPatientTitle = async (req, res) => {
+  const permission = getGrants.can(req.user.role).readAny("patientTitle");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden " + req.user.role);
+  }
+
   await db.all("SELECT * FROM PatientTitle", function (err, rows) {
     return res.json(rows);
   });
 };
 
 exports.updatePatientTitle = async (req, res) => {
+
+  const permission = getGrants.can(req.user.role).updateAny("patientTitle");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden " + req.user.role);
+  }
+
   await db.run("Update PatientTitle set title=? where id=?", [req.body.title, req.params["id"]], function (err) {
     if (err) {
       return res.status(400).send(err.message);
